backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
motor==3.3.2
pymongo==4.6.1
pydantic==2.6.3
pydantic-settings==2.2.1
python-dotenv==1.0.1
websockets==12.0
httpx==0.27.0
